pyproject.toml
[project]
name = "urbtec-utility-labelling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "streamlit-cookies-controller>=0.0.4",
    "streamlit-drawable-canvas>=0.9.3",
    "watchdog>=6.0.0",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
search-path = ["./", "./src", "./src/api/routes/tasks/models.py"]
permissive-ignores = true
